requirements.txt
beautifulsoup4==4.4.1
NetEase-MusicBox==0.2.2.10
nose==1.3.7
numpy==1.11.1
Pillow==3.3.1
powerline-status==2.4
pycrypto==2.6.1
PyQt5==5.7
pytesseract==0.1.6
requests==2.10.0
sip==4.18.1
virtualenv==15.0.2
